# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Optimax HR & Internship Management Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PaymentPaymentStatus = typeof PaymentPaymentStatus[keyof typeof PaymentPaymentStatus];


export const PaymentPaymentStatus = {
  paid: 'paid',
  partial: 'partial',
  pending: 'pending',
} as const;

export interface Payment {
  id: string;
  student_id: string;
  total_fee: number;
  amount_paid: number;
  balance_amount: number;
  payment_status: PaymentPaymentStatus;
  /** @nullable */
  updated_at?: string | null;
}

export interface Student {
  id: string;
  full_name: string;
  email: string;
  /** @nullable */
  phone_number?: string | null;
  internship_role: string;
  start_date: string;
  end_date: string;
  /** @nullable */
  created_by?: string | null;
  created_at: string;
  payment?: Payment;
}

export interface StudentInput {
  full_name: string;
  email: string;
  phone_number?: string;
  internship_role: string;
  start_date: string;
  end_date: string;
  total_fee?: number;
  amount_paid?: number;
}

export interface StudentUpdate {
  full_name?: string;
  email?: string;
  phone_number?: string;
  internship_role?: string;
  start_date?: string;
  end_date?: string;
}

export interface StudentList {
  data: Student[];
  total: number;
  page: number;
  limit: number;
}

export interface PaymentUpdate {
  total_fee?: number;
  amount_paid?: number;
}

export type OfferLetterStatus = typeof OfferLetterStatus[keyof typeof OfferLetterStatus];


export const OfferLetterStatus = {
  generated: 'generated',
  not_generated: 'not_generated',
} as const;

export interface OfferLetter {
  id: string;
  student_id: string;
  generated_at: string;
  /** @nullable */
  file_url?: string | null;
  status: OfferLetterStatus;
  student?: Student;
}

export interface OfferLetterInput {
  student_id: string;
  joining_date?: string;
}

export interface OfferLetterUpdate {
  status?: string;
  file_url?: string;
}

export type CertificateStatus = typeof CertificateStatus[keyof typeof CertificateStatus];


export const CertificateStatus = {
  issued: 'issued',
  not_issued: 'not_issued',
} as const;

export interface Certificate {
  id: string;
  student_id: string;
  generated_at: string;
  /** @nullable */
  file_url?: string | null;
  status: CertificateStatus;
  student?: Student;
}

export interface CertificateInput {
  student_id: string;
  issue_date?: string;
}

export interface CertificateUpdate {
  status?: string;
  file_url?: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  hr: 'hr',
} as const;

export interface User {
  id: string;
  name: string;
  email: string;
  role: UserRole;
  created_at: string;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  admin: 'admin',
  hr: 'hr',
} as const;

export interface UserInput {
  name: string;
  email: string;
  role: UserInputRole;
  password?: string;
}

export interface UserUpdate {
  name?: string;
  email?: string;
  role?: string;
}

export interface HrSummary {
  total_students: number;
  offer_letters_generated: number;
  certificates_issued: number;
  pending_payments: number;
  partial_payments?: number;
}

export interface AdminSummary {
  total_students: number;
  total_applications?: number;
  total_revenue: number;
  paid_revenue: number;
  pending_revenue: number;
  offer_letters_generated: number;
  certificates_issued: number;
}

export interface ActivityEvent {
  id: string;
  type: string;
  description: string;
  timestamp: string;
  /** @nullable */
  student_name?: string | null;
}

export interface MonthlyRevenue {
  month: string;
  total: number;
  paid: number;
  pending: number;
}

export interface PaymentBreakdown {
  paid: number;
  partial: number;
  pending: number;
}

export interface StudentGrowth {
  month: string;
  count: number;
}

export type ListStudentsParams = {
search?: string;
status?: string;
role?: string;
page?: number;
limit?: number;
};

export type ListPaymentsParams = {
status?: string;
student_id?: string;
};

export type ListOfferLettersParams = {
status?: string;
student_id?: string;
};

export type ListCertificatesParams = {
status?: string;
student_id?: string;
};

export type GetRecentActivityParams = {
limit?: number;
};