# lib/api-zod/src/generated/types/offerLetterUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Optimax HR & Internship Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface OfferLetterUpdate {
  status?: string;
  file_url?: string;
}