# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Optimax HR & Internship Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Payment } from './payment';

export interface Student {
  id: string;
  full_name: string;
  email: string;
  /** @nullable */
  phone_number?: string | null;
  internship_role: string;
  start_date: string;
  end_date: string;
  /** @nullable */
  created_by?: string | null;
  created_at: string;
  payment?: Payment;
}