# lib/api-zod/src/generated/types/studentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Optimax HR & Internship Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface StudentInput {
  full_name: string;
  email: string;
  phone_number?: string;
  internship_role: string;
  start_date: string;
  end_date: string;
  total_fee?: number;
  amount_paid?: number;
}